Forward all metadata, always
There was a bug where int type variables didn't have their descriptions
being forwarded. This bug was caused by two independent bugs.

The first one being that the "all" category of metadata tags was only
ever being forwarded, if the type of a variable was recognized to be
inside metadata_aliases. Requiring the type to be recognized
preemptively defeats the purpose of the "all" list of metadatatags to
begin with.

Second, ints were being resolved to a type named "integer", and floats
were being resolved to a type named "number". "number" was the only
numeric type to be recognized, which appears to be another bug. I've
made "int" resolve to "number" as well.