Add self to Pipfile
In order to run the example, we need to have this project itself to be
an added dependency. The Pipfile needs to be cleaned up, such that we
can separate the dependencies for running the example, and the
dependencies that are actually required for this project.

I'll do that eventually.