Replace response.model with cattrs
Cattrs is a library which works with attrs data to SerDe the classes to
dictionaries. I already like using cattrs elsewhere, and the ability to
undo some of these classes doesn't feel like it belongs in this library
anyways. We're only using cattrs in the sample code as it currently
stands.