Add tests to cover >90% of the project
Thankfully, the code base is really small, so it was easy to cover all
of the existing code in tests.

I did some small refactoring in the codebase as well. Mostly renaming
functions or variables that I thought were poorly named.

The code itself doesn't have any functional changes in it. I've just
added a bunch of tests to verify the current behaviors of the code. I
can re-work the code after this.