Remove large license section, replace with badge
The FOSSA license section that got added is very slow to load, and since
we aren't not using the paid version of FOSSA, the pane doesn't show
that much useful information. Instead, I'm just replacing it with a
badge for the MIT license, that should make things a little more clear.