Make uploading to codecov a different step
Uploading to codecov requires a token, but we don't want to expose that
token to just any person who submits a PR. Otherwise, someone could
submit a PR which simply runs `echo $CODECOV_TOKEN` and expose the
token.

So I've split out the build step in CircleCI which uploads to codecov,
and we'll only run it after things get merged. We still have the minimum
coverage percentage. I'll just make sure that this step which uses the
token will not be a part of merging a PR.